consolidate auto-assign vpc ipv4 address logic into one component
