+++ packages/manager/src/features/Linodes/LinodeCreate/Networking/VPC.tsx
@@ -7,23 +7,19 @@ import {
   FormControlLabel,
   Notice,
   Stack,
-  TextField,
-  TooltipIcon,
   Typography,
 } from '@linode/ui';
 import { LinkButton } from '@linode/ui';
 import React, { useState } from 'react';
 import { Controller, useFormContext, useWatch } from 'react-hook-form';
 
+import { VPCIPv4Address } from 'src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/VPCIPv4Address';
 import { VPCIPv6Address } from 'src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/VPCIPv6Address';
 import {
   VPCIPv6PublicIPLabel,
   VPCPublicIPLabel,
 } from 'src/features/VPCs/components/VPCPublicIPLabel';
-import {
-  REGION_CAVEAT_HELPER_TEXT,
-  VPC_AUTO_ASSIGN_IPV4_TOOLTIP,
-} from 'src/features/VPCs/constants';
+import { REGION_CAVEAT_HELPER_TEXT } from 'src/features/VPCs/constants';
 import { generateVPCIPv6InputHelperText } from 'src/features/VPCs/utils';
 import { VPCCreateDrawer } from 'src/features/VPCs/VPCCreateDrawer/VPCCreateDrawer';
 import { useVPCDualStack } from 'src/hooks/useVPCDualStack';
@@ -160,41 +156,20 @@ export const VPC = ({ index }: Props) => {
             control={control}
             name={`linodeInterfaces.${index}.vpc.ipv4.addresses.0.address`}
             render={({ field, fieldState }) => (
-              <Box>
-                <FormControlLabel
-                  checked={field.value === 'auto'}
-                  control={<Checkbox sx={{ ml: 0.4 }} />}
-                  disabled={!regionSupportsVPCs}
-                  label={
-                    <Stack alignItems="center" direction="row">
-                      <Typography>Auto-assign VPC IPv4 address</Typography>
-                      <TooltipIcon
-                        status="info"
-                        text={VPC_AUTO_ASSIGN_IPV4_TOOLTIP}
-                      />
-                    </Stack>
-                  }
-                  onChange={(e, checked) =>
-                    field.onChange(checked ? 'auto' : '')
-                  }
-                />
-                {field.value !== 'auto' && (
-                  <TextField
-                    containerProps={{ sx: { mb: 1.5, mt: 1 } }}
-                    errorText={
-                      fieldState.error?.message ??
-                      errors.linodeInterfaces?.[index]?.vpc?.ipv4
-                        ?.addresses?.[0]?.message
-                    }
-                    label="VPC IPv4"
-                    noMarginTop
-                    onBlur={field.onBlur}
-                    onChange={field.onChange}
-                    required
-                    value={field.value}
-                  />
+              <VPCIPv4Address
+                disabled={!regionSupportsVPCs}
+                errorMessage={
+                  fieldState.error?.message ??
+                  errors.linodeInterfaces?.[index]?.vpc?.ipv4?.addresses?.[0]
+                    ?.message
+                }
+                fieldValue={field.value}
+                helperText={generateVPCIPv6InputHelperText(
+                  selectedSubnet?.ipv6?.[0].range ?? ''
                 )}
-              </Box>
+                onBlur={field.onBlur}
+                onChange={field.onChange}
+              />
             )}
           />
           {showIPv6Fields && (
