--- packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/AddInterfaceDrawer/VPC/VPCIPv4Address.tsx
@@ -1,84 +1,0 @@
-import {
-  Checkbox,
-  FormControlLabel,
-  Notice,
-  Stack,
-  TextField,
-  TooltipIcon,
-  Typography,
-} from '@linode/ui';
-import React from 'react';
-import { Controller, useFormContext } from 'react-hook-form';
-
-import { ErrorMessage } from 'src/components/ErrorMessage';
-import { VPC_AUTO_ASSIGN_IPV4_TOOLTIP } from 'src/features/VPCs/constants';
-
-import type { CreateInterfaceFormValues } from '../utilities';
-
-interface Props {
-  index: number;
-  isDualStackVPC: boolean;
-}
-
-export const VPCIPv4Address = (props: Props) => {
-  const { index, isDualStackVPC } = props;
-  const {
-    control,
-    formState: { errors },
-  } = useFormContext<CreateInterfaceFormValues>();
-
-  const error = errors.vpc?.ipv4?.addresses?.[index]?.message;
-
-  return (
-    <Stack spacing={1}>
-      {error && (
-        <Notice variant="error">
-          <ErrorMessage message={error} />
-        </Notice>
-      )}
-      <Controller
-        control={control}
-        name={`vpc.ipv4.addresses.${index}.address`}
-        render={({ field, fieldState }) => (
-          <Stack rowGap={1}>
-            <Stack direction="row">
-              <FormControlLabel
-                checked={field.value === 'auto'}
-                control={<Checkbox />}
-                label="Auto-assign VPC IPv4 address"
-                onChange={(e, checked) => field.onChange(checked ? 'auto' : '')}
-                sx={{ pl: 0.4, mr: 0 }}
-              />
-              <TooltipIcon
-                status="info"
-                text={
-                  isDualStackVPC ? (
-                    <Typography component="span">
-                      Automatically assign an IPv4 address as{' '}
-                      {isDualStackVPC ? 'a' : 'the'} private IP address for this
-                      Linode in the VPC.
-                    </Typography>
-                  ) : (
-                    VPC_AUTO_ASSIGN_IPV4_TOOLTIP
-                  )
-                }
-              />
-            </Stack>
-            {field.value !== 'auto' && (
-              <TextField
-                containerProps={{ sx: { mb: 1.5, mt: 1 } }}
-                errorText={fieldState.error?.message}
-                label="VPC IPv4"
-                noMarginTop
-                onBlur={field.onBlur}
-                onChange={field.onChange}
-                required
-                value={field.value}
-              />
-            )}
-          </Stack>
-        )}
-      />
-    </Stack>
-  );
-};
