+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/LinodeInterfaces/EditInterfaceDrawer/VPCInterface/VPCIPAddresses.tsx
@@ -5,8 +5,8 @@ import { useFieldArray, useFormContext, useWatch } from 'react-hook-form';
 import { ErrorMessage } from 'src/components/ErrorMessage';
 import { useVPCDualStack } from 'src/hooks/useVPCDualStack';
 
+import { EditVPCIPv4Address } from './EditVPCIPv4Address';
 import { EditVPCIPv6Address } from './EditVPCIPv6Address';
-import { VPCIPv4Address } from './VPCIPv4Address';
 
 import type {
   LinodeInterface,
@@ -48,7 +48,7 @@ export const VPCIPAddresses = (props: Props) => {
         </Notice>
       )}
       {fields.map((field, index) => (
-        <VPCIPv4Address
+        <EditVPCIPv4Address
           index={index}
           isDualStackVPC={isDualStackVPC}
           key={field.id}
